requirements_cluster.txt
absl-py==1.4.0
accelerate==0.28.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bert-score==0.3.13
bitsandbytes==0.43.0
bleach==6.1.0
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
coloredlogs==15.0.1
comm==0.1.4
confection==0.1.1
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.7
dataclasses-json==0.6.1
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.12.3
defusedxml==0.7.1
dill==0.3.7
distro==1.8.0
docstring_parser==0.16
einops==0.7.0
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.9.0
flash-attn==2.3.2
fonttools==4.45.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.34
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.57.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.21.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
interegular==0.3.2
intertext-graph @ git+https://github.com/UKPLab/intertext-graph.git@2516bd20a7153e825b89206dac0ebeb7d1ab7302
ipdb==0.13.13
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
langchain==0.0.304
langcodes==3.3.0
langsmith==0.0.41
lightning-utilities==0.9.0
lit==15.0.7
lm-format-enforcer==0.7.2
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpi4py @ file:///work/ci_py311/mpi4py_1676858691457/work
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
ninja==1.11.1.1
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.3.2
optimum==1.13.2
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
peft==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.11.0
preshed==3.0.8
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.2.0
pydantic_core==2.6.0
pydevd-pycharm==232.9559.58
Pygments==2.16.1
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.1.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
rank-bm25==0.2.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.10.4
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.11.4
Send2Trash==1.8.2
sentence-transformers==2.7.0
sentencepiece==0.1.99
shtab==1.7.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.21
srsly==2.4.7
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.17.1
thinc==8.1.12
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.15.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.0.3
torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.38.2
triton==2.0.0
trl==0.8.1
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tyro==0.7.3
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.13
wasabi==1.1.2
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.3.7
widgetsnbextension==4.0.9
xxhash==3.4.1
yarl==1.9.2
